Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Let me explain what this does:
The FFC is only VGA but the original HAL has a default video mode of 1920x1080 causing the FFC to crash. I had set the default mode to VGA and this worked for everything except apps that capture preview frames because they get a resolution they aren't expecting. This fix is a bit of a hack because it only sets the default mode to VGA if autofocus is disabled, i.e for the front cam. There's probably a better way of doing this but I couldn't think of one at the time.

Patch-set: 1