Update patch set 1
Patch Set 1:

I'm surprised ZSL works at all, is it even useful since it's restricted to <= XGA res? There are lots of checks for ZSL mode in the code, I'll take a look to see why it would fail to  encode thumbnails but perhaps it's best to keep it disabled. The problem might be in the proprietary code anyway, does HTC use this mode?

Patch-set: 1