Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This fixes the corrupt thumbnail but the flash still won't work. I think we need a camera setting to enable/disable ZSL mode for QCOM hardware. The other option is to enable it only if the flash is off.

Patch-set: 1