Update patch set 1
Patch Set 1:

not sure if you saw this on git, but heres a copy and paste.
Hey dzo-
Some information reguarding zsl. after speaking to a sammy cm10 user, we were able to figure out the logic behind them using zsl.
if flash is set to auto - if the picture is bright enough without the need for flash, it snaps a picture instantly using zsl. if the room is dark where it needs flash, then it is bypassing zsl and snapping the picture normally with flash. 
if flash is set to on - zsl mode is never enabled. 
If flash is set to off - zsl mode is always enabled.
I hope this helps out. Thanks for all that you have done.
-deck

Patch-set: 1