Update patch set 1
Patch Set 1:

well i thought that was gonna fix our voip problem, but  it turns out i was wrong. That was the real reason changed, tho everywhere else i looked had it the same too. iunno, need to look into voip more. Need to dig and see if those bits are actually supported or not.

Patch-set: 1