Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Without this, compile breaks. The commit that breaks compile is needed, and has tried to be pushed in about 4 times. We must act in the device trees since that commit will not be reverted.

I would at least like this to be implemented temporarily until we update display goodies in kernel.

Patch-set: 1