Update patch set 1
Patch Set 1:

This adds the option in Settings -> System -> Power menu -> Expanded desktop, which when enabled adds to the Power menu an toggle which will hide/unhide the status bar and software buttons (so you can have the full screen on apps that don't already have this functionality baked in).

This is more useful on phones/tablets with software buttons, but I figured it couldn't hurt to have the option added (it isn't enabled by default).

Patch-set: 1