Update patch set 1
Patch Set 1: (1 inline comment)

Worked perfectly after 10 reboots for me. Works as we wanted it to as far as Evita is concerned.

We might consider leaving old RIL for now. Would hate to have to go through the pain and wait to get it re-added later...

Doing patchset to fix my alphabetical mistake :P.

Patch-set: 1