Update patch set 1
Patch Set 1:

Well, for one the kernel doesn't have the low latency bits (no device, no msm-lowlatency-pcm-q6.c, etc). I checked, and the only reason it seems to work is because it errors out early when it fails to find the right pcm. (snd_soc_msm sets it to 14, which is bogus, that's an FM pcm, and not active)

Patch-set: 1