Update patch set 1
Patch Set 1:

because this stuff depends on the kernel used.
example: if we upgrade m7 kernel you have to move it back down to s4 common since m7 doesn't need it anymore then. ;-)

the only things which are really common is stuff which never changes. for example cpu arch arm and so on.

Patch-set: 1