Update patch set 1
Patch Set 1:

I would wait till we had a look at the google edition. Currently nobody know's what they did and how it will look like.

We played the "move it up, move it down"-game on the sony devices because of to much "common" and the same will probably happen here. It's a waste of time.

And if HTC is going to release a tablet soon you'll have to move the phone specific stuff down anyway.

Patch-set: 1