Update patch set 1
Patch Set 1:

This one. It seems like any adjustment of the sample rate is enough to desync a2dp. I suspect it has something to do with the AFE proxy audio device that is used, but I'm not completely sure. We don't see this on other devices though.

This patch had a derp (corrected in the followup patch), but simply keeping it the same was enough to stop the crazy audio stuttering.

Patch-set: 1