Update patch set 1
Patch Set 1:

cellinfolist isn't supported by a lot of older radio libs, there's some provision to disable that in libril using a NO_CELLINFOLIST flag. I haven't, but I've been running it on m7 for a day or so. Wakelocks were an issue last time I had to deal with qcril, so that's mainly why I asked.

Patch-set: 1