Update patch set 1
Patch Set 1: Code-Review-2

Let's not override the network_mode list for no real reason. I'm assuming this is to hide the TD-SCDMA options from the *#*#INFO#*#* menu? If that's the case, we have always shown unsupported network modes there. Additionally, that would be considered a debug menu, so it's not a big deal.

Patch-set: 1