Update patch set 1
Patch Set 1: -Code-Review

I'm kind of out of the loop on this audio stuff. What's the end goal here? Use awesomeplayer or nuplayer? I think I recall effects not working for tunnel audio, is this also the case for LPA? Are we basically disabling all audio hotness in favor of "it just works" with this commit? Will my string of questions ever end?

Patch-set: 1