Update patch set 1
Patch Set 1: Code-Review+1

Not sure how this is a hack exactly.  I created this as inspired by Ethan's recommendation to look at a motorola device, then this was taken from my inspiration.  The proximity sensor is a wake sensor, so no additional battery drain should occur.  We're simply applying some action to the proximity sensor broadcasts.

I'm not saying that just because other devices have it, that it's right, but it's almost not fair to exclude from other devices that'd like to implement it.

Patch-set: 1
Label: Code-Review=+1