Update patch set 2
Patch Set 2:

Proposed alternative to http://review.cyanogenmod.org/169122 . Instead of dropping the response of solicited request RIL_REQUEST_VOICE_RADIO_TECH, just ensure it gets set to CDMA type.

I'm not terribly pleased with this, as it means users with GSM sims in these devices will need to remove telephony.lteOnCdmaDevice from build.prop, but then, this puts us par for the course with CM11.0 and CM12.1.

Thoughts on whether this approach is acceptable?

Patch-set: 2