msm8960-common: Bump RIL version
The RIL versions reported by RIL_register span from 6 (fireball) - 11
(m7_ul) for the devices supported by this -common configuration. For
the purposes of building libril, there is no difference in setting
RIL_VERSION and LAST_IMPRECISE_RIL_VERSION to any of 10, 11, or 12, so
long as the version is greater or equal to that of the vendor provided
rild. Set version 11 for all devices.

Change-Id: I7c6dfc433eff811520e56f4934171fe51aa0f9ac