Revert "msm8960-common: Use RIL version 10"
The RIL versions reported by RIL_register span from 6 (fireball) - 11
(m7_ul) for the devices supported by this -common configuration. For
the purposes of building libril, there is no difference in setting
RIL_VERSION and LAST_IMPRECISE_RIL_VERSION to any of 10, 11, or 12, so
long as the version is greater or equal to that of the vendor provided
rild. Accept the default of 12.

This reverts commit f67acf95b219662e407ae7018d5bb95f65e42382.

Change-Id: I7c6dfc433eff811520e56f4934171fe51aa0f9ac